backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.28.1
supabase==2.20.0
python-dateutil==2.9.0
gunicorn==21.2.0
